+++ packages/network-controller/src/NetworkController.ts
@@ -40,6 +40,7 @@ import { v4 as uuidV4 } from 'uuid';
 import {
   DEPRECATED_NETWORKS,
   INFURA_BLOCKED_KEY,
+  FALLBACK_INFURA_NETWORK_TYPE_MAPPING,
   NetworkStatus,
 } from './constants';
 import type {
@@ -823,9 +824,28 @@ function getDefaultInfuraNetworkConfigurationsByChainId(): Record<
   Hex,
   NetworkConfiguration
 > {
+  const reverseFallbackInfuraNetworkTypeMapping = Object.entries(
+    FALLBACK_INFURA_NETWORK_TYPE_MAPPING,
+  ).reduce<Record<InfuraNetworkType, string>>(
+    (reversed, [key, value]) => {
+      reversed[value] = key;
+      return reversed;
+    },
+    {} as Record<InfuraNetworkType, string>,
+  );
+
   return Object.values(InfuraNetworkType).reduce<
     Record<Hex, NetworkConfiguration>
-  >((obj, infuraNetworkType) => {
+  >((obj, _infuraNetworkType) => {
+    let infuraNetworkType = _infuraNetworkType;
+    if (
+      hasProperty(reverseFallbackInfuraNetworkTypeMapping, infuraNetworkType)
+    ) {
+      infuraNetworkType = reverseFallbackInfuraNetworkTypeMapping[
+        infuraNetworkType
+      ] as InfuraNetworkType;
+    }
+
     const chainId = ChainId[infuraNetworkType];
 
     // Skip deprecated network as default network.
@@ -845,7 +865,7 @@ function getDefaultInfuraNetworkConfigurationsByChainId(): Record<
       rpcEndpoints: [
         {
           failoverUrls: [],
-          networkClientId: infuraNetworkType,
+          networkClientId: _infuraNetworkType,
           type: RpcEndpointType.Infura,
           url: rpcEndpointUrl,
         },
@@ -1116,6 +1136,18 @@ function deriveInfuraNetworkNameFromRpcEndpointUrl(
   if (match?.groups) {
     if (isInfuraNetworkType(match.groups.networkName)) {
       return match.groups.networkName;
+    }
+
+    // Infura URL can be inserted by user manually,
+    // in case the network name extra from URL is not mapped to InfuraNetworkType,
+    // we should add a fallback mapping here.
+    if (
+      hasProperty(
+        FALLBACK_INFURA_NETWORK_TYPE_MAPPING,
+        match.groups.networkName,
+      )
+    ) {
+      return FALLBACK_INFURA_NETWORK_TYPE_MAPPING[match.groups.networkName];
     }
 
     throw new Error(`Unknown Infura network '${match.groups.networkName}'`);